requirements.txt
streamlit==1.38.0
langchain==0.3.7
langchain-community==0.3.7
langchain-google-genai==2.0.4
google-generativeai>=0.8.0,<0.9.0
faiss-cpu==1.8.0.post1
pypdf==4.3.1
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.30.6
tiktoken==0.7.0
